servlet/src/main/java/tutorial/servlet/domain/member/MemberRepository.java
package tutorial.servlet.domain.member;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

/*
동시성 문제가 고려되지 않는 HashMap 대신 ConcurrentHashMap, AtomicLong을 고려해야하지만 일단 이렇게함
* */
public class MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    private static final MemberRepository instance = new MemberRepository();
    //spring을 사용하지 않고 있어서 직접 싱글톤으로 만들어야함

    public static MemberRepository getInstance(){
        return instance;
    }

    private MemberRepository(){

    }

    public Member save(Member member){
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    public Member findById(Long id){
        return store.get(id);
    }

    public List<Member> findAll(){
        return new ArrayList<>(store.values());
    }

    //test용
    public void clearStore(){
        store.clear();
    }




}
